+++ network/tor/tor/build.gradle.kts
@@ -31,6 +31,7 @@ dependencies {
     implementation(project(":tor-common"))
 
     implementation("bisq:security")
+    implementation("bisq:java-se")
 
     implementation("network:network-identity")
     implementation("network:socks5-socket-channel")
